add update and filter by date

+++ expenses_app/views.py
@@ -5,10 +5,20 @@ from django.shortcuts import get_object_or_404
 from .models import Expense
 from .serializers import ExpenseSerializer
 
+
 @api_view(['GET', 'POST'])
 def expense_list(request):
     if request.method == 'GET':
         expenses = Expense.objects.all().order_by('-date')
+
+        category = request.GET.get('category')
+        date = request.GET.get('date')
+
+        if category:
+            expenses = expenses.filter(category=category)
+        if date:
+            expenses = expenses.filter(date=date)
+
         serializer = ExpenseSerializer(expenses, many=True)
         return Response(serializer.data)
 
@@ -19,8 +29,22 @@ def expense_list(request):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['DELETE'])
+
+@api_view(['GET', 'PUT', 'DELETE'])
 def expense_detail(request, pk):
     expense = get_object_or_404(Expense, pk=pk)
-    expense.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+
+    if request.method == 'GET':
+        serializer = ExpenseSerializer(expense)
+        return Response(serializer.data)
+
+    elif request.method == 'PUT':
+        serializer = ExpenseSerializer(expense, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    elif request.method == 'DELETE':
+        expense.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
